rename: 플레이리스트 정보(생성, 수정) 관련 api파일명 변경 (#116)
